+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to
 - 💄(frontend) error alert closeable on editor #284
 - ♻️(backend) Change email content #283
 - 🛂(frontend) viewers and editors can access share modal #302
+- ♻️(frontend) remove footer on doc editor #313
 
 ## Fixed